Added temperature degrade, if el price is very high.

+++ main/display.h
@@ -2,7 +2,7 @@
 #define __DISPLAY__
 
 int display_init(void);
-void display_show(float f1, float f2);
+void display_show(int f1, float f2);
 void display_close(void);
 void display_text(char *t);
 void display_clear(void);
